+++ changes.md
@@ -31,7 +31,7 @@
 26. Draw X/Y/Z axis titles in lego plots
 27. Use canvas Theta/Phi angles to set initial camera position in 3D plots 
 28. Create statbox when drawing TGraph with TF1 in list of functions
-29. Implement drawing of TSpline3 class 
+29. Implement drawing of TSpline3 and TSpline5 classes 
 
 
 ## Changes in 5.2.0